+++ server/src/server.js
@@ -2,8 +2,8 @@ const express = require('express');
 const path = require('path');
 const db = require('./config/connection');
 const { ApolloServer } = require('apollo-server-express');
-const { typeDefs, resolvers } = require('./schema');
-const { authMiddleware } = require('./utils/auth');
+const { typeDefs, resolvers } = require('../schema');
+const { authMiddleware } = require('../utils/auth');
 
 const app = express();
 const PORT = process.env.PORT || 3001;
